Reorganized Project. WIP Heap::Solve general solver function.
Heap::Solve is currently very buggy, crashing upon room unloading when calling Deallocate, and I'm not sure why. The rest of the solver functions seem to be stable though even with the project reorganization.

+++ Project Oceanside/include/Scene.h
@@ -14,7 +14,7 @@ public:
 	void DumpSceneInfo() const;
 	void ResetClearedActors();
 	std::map<int, Node*> GetTransitionActors() const;
-	int NumberOfTransitionActors() const;
+	size_t NumberOfTransitionActors() const;
 
 private:
 	void LoadScene();
